// packages/asistencia/src/employees/dto/create-empleado.dto.d.ts
import { Planta } from '../../entities/empleado.entity';
export declare class CreateEmpleadoDto {
    firstName: string;
    lastName: string;
    dni?: string;
    pin: string;
    planta: Planta;
    departamento?: string;
    cargo?: string;
    activo?: boolean;
}
//# sourceMappingURL=create-empleado.dto.d.ts.map